[SPARK-56425] Split `DataFrame.swift` into extension-based files
